requirements.txt
coqui-tts
python-telegram-bot

